+++ src/main/java/net/contargo/print/pdf/QRGenRenderer.java
@@ -1,7 +1,9 @@
 package net.contargo.print.pdf;
 
+import com.google.zxing.EncodeHintType;
 import com.google.zxing.qrcode.decoder.ErrorCorrectionLevel;
 
+import net.glxn.qrgen.core.AbstractQRCode;
 import net.glxn.qrgen.core.exception.QRGenerationException;
 import net.glxn.qrgen.javase.QRCode;
 
@@ -15,21 +17,18 @@ import net.glxn.qrgen.javase.QRCode;
 public class QRGenRenderer implements QRCodeRenderer {
 
     @Override
-    public byte[] render(String code, int size) throws RenderException {
-
-        return render(code, size, -1);
-    }
-
-
-    @Override
-    public byte[] render(String code, int size, int level) throws RenderException {
+    public byte[] render(String code, int size, int level, boolean margin) throws RenderException {
 
         try {
-            return QRCode.from(code)
-                .withErrorCorrection(toErrorCorrectionLevel(level))
-                .withSize(size, size)
-                .stream()
-                .toByteArray();
+            AbstractQRCode c = QRCode.from(code)
+                    .withErrorCorrection(toErrorCorrectionLevel(level))
+                    .withSize(size, size);
+
+            if (!margin) {
+                c = c.withHint(EncodeHintType.MARGIN, 0);
+            }
+
+            return c.stream().toByteArray();
         } catch (QRGenerationException e) {
             throw new RenderException("QR-code render failed.", e);
         }